Dialog : introduce DialogLocker
