+++ vector/src/main/java/im/vector/dialogs/ConsentNotGivenHelper.kt
@@ -17,8 +17,10 @@
 package im.vector.dialogs
 
 import android.app.Activity
+import android.app.AlertDialog
 import android.os.Bundle
-import android.support.v7.app.AlertDialog
+import android.text.method.LinkMovementMethod
+import android.widget.TextView
 import im.vector.Matrix
 import im.vector.R
 import im.vector.activity.VectorWebViewActivity
@@ -27,15 +29,13 @@ import im.vector.webview.WebViewMode
 import org.matrix.androidsdk.rest.model.MatrixError
 import org.matrix.androidsdk.util.Log
 
-class ConsentNotGivenHelper(private val activity: Activity, savedInstanceState: Bundle?) :
-        Restorable {
+private const val LOG_TAG = "ConsentNotGivenHelper"
 
-    /* ==========================================================================================
-     * Data
-     * ========================================================================================== */
+class ConsentNotGivenHelper private constructor(private val activity: Activity,
+                                                private val dialogLocker: DialogLocker) :
+        Restorable by dialogLocker {
 
-    // Ensure the dialog is not displayed multiple times
-    private var isDialogDisplayed = savedInstanceState?.getBoolean(KEY_DIALOG_IS_DISPLAYED, false) == true
+    constructor(activity: Activity, savedInstanceState: Bundle?) : this(activity, DialogLocker(savedInstanceState))
 
     /* ==========================================================================================
      * Public methods
@@ -45,59 +45,27 @@ class ConsentNotGivenHelper(private val activity: Activity, savedInstanceState: 
      * Display the consent dialog, if not already displayed
      */
     fun displayDialog(matrixError: MatrixError) {
-        if (isDialogDisplayed) {
-            // Filter this request
-            Log.w(LOG_TAG, "Filtered dialog request")
-            return
-        }
-
-        // Check required parameter
         if (matrixError.consentUri == null) {
             Log.e(LOG_TAG, "Missing required parameter 'consent_uri'")
             return
         }
-
-        isDialogDisplayed = true
-
-        AlertDialog.Builder(activity)
-                .setTitle(R.string.settings_app_term_conditions)
-                .setMessage(activity.getString(R.string.dialog_user_consent_content,
-                        Matrix.getInstance(activity).defaultSession.homeServerConfig.homeserverUri.host))
-                .setPositiveButton(R.string.dialog_user_consent_submit) { _, _ ->
-                    openWebViewActivity(matrixError.consentUri)
-                    isDialogDisplayed = false
-                }
-                .setNegativeButton(R.string.later) { _, _ ->
-                    isDialogDisplayed = false
-                }
-                .setOnCancelListener { isDialogDisplayed = false }
-                .show()
+        dialogLocker.displayDialog {
+            AlertDialog.Builder(activity)
+                    .setTitle(R.string.settings_app_term_conditions)
+                    .setMessage(activity.getString(R.string.dialog_user_consent_content,
+                            Matrix.getInstance(activity).defaultSession.homeServerConfig.homeserverUri.host))
+                    .setPositiveButton(R.string.dialog_user_consent_submit) { _, _ ->
+                        openWebViewActivity(matrixError.consentUri)
+                    }
+        }
     }
 
-    /* ==========================================================================================
-     * Implements Restorable
-     * ========================================================================================== */
-
-    override fun saveState(outState: Bundle) {
-        outState.putBoolean(KEY_DIALOG_IS_DISPLAYED, isDialogDisplayed)
-    }
-
-    /* ==========================================================================================
-     * Private
-     * ========================================================================================== */
+/* ==========================================================================================
+* Private
+* ========================================================================================== */
 
     private fun openWebViewActivity(consentUri: String) {
         val intent = VectorWebViewActivity.getIntent(activity, consentUri, R.string.settings_app_term_conditions, WebViewMode.CONSENT)
         activity.startActivity(intent)
-    }
-
-    /* ==========================================================================================
-     * COMPANION
-     * ========================================================================================== */
-
-    companion object {
-        private const val LOG_TAG = "ConsentNotGivenHelper"
-
-        private const val KEY_DIALOG_IS_DISPLAYED = "ConsentNotGivenHelper.KEY_DIALOG_IS_DISPLAYED"
     }
 }
